requirements.txt
rich>=13.0.0
questionary>=2.0.0
openai>=1.0.0
pyyaml>=6.0
gradio>=4.0.0
